fix/Entiy spot and event for Type Float Longitude and Latitude

+++ high-riders/src/Entity/Spot.php
@@ -71,7 +71,7 @@ class Spot
     private $city;
 
     /**
-     * @ORM\Column(type="integer", nullable=true)
+     * @ORM\Column(type="float", nullable=true)
      * 
      * 
      * @Groups({"spot_list", "spot_detail"})
@@ -80,7 +80,7 @@ class Spot
     private $longitude;
 
     /**
-     * @ORM\Column(type="integer", nullable=true)
+     * @ORM\Column(type="float", nullable=true)
      * 
      * 
      * @Groups({"spot_list", "spot_detail"})
@@ -352,24 +352,24 @@ class Spot
         return $this;
     }
 
-    public function getLongitude(): ?int
+    public function getLongitude(): ?float
     {
         return $this->longitude;
     }
 
-    public function setLongitude(?int $longitude): self
+    public function setLongitude(?float $longitude): self
     {
         $this->longitude = $longitude;
 
         return $this;
     }
 
-    public function getLatitude(): ?int
+    public function getLatitude(): ?float
     {
         return $this->latitude;
     }
 
-    public function setLatitude(?int $latitude): self
+    public function setLatitude(?float $latitude): self
     {
         $this->latitude = $latitude;
 
@@ -702,5 +702,5 @@ class Spot
 
         return $this;
     }
-    
+
 }
